+++ spectrum_overload/test/test_DifferentialSpectrum.py
@@ -1,4 +1,7 @@
+# -*- coding: utf-8 -*-
+
 """Test DifferentialSpectrum Class."""
+from __future__ import division, print_function
 
 import numpy as np
 import pytest